# docs/swagger.yaml
definitions:
...
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.CategoryRes:
    properties:
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.ChatMessageRes:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      sender:
        type: string
    type: object
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.ChatSessionRes:
    properties:
      contextSummary:
        type: string
      createdAt:
        type: string
      id:
        type: string
      isArchived:
        type: boolean
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.CreateChatSessionReq:
    properties:
      title:
        type: string
    type: object
...
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.RecommendOutfitReq:
    properties:
      details:
        type: string
      occasion:
        type: string
      season:
        type: string
      styleTarget:
        type: string
      weather:
        type: string
    type: object
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.RecommendedItemGroup:
    properties:
      alternatives:
        items:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.WardrobeItemRes'
        type: array
      primary:
        $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.WardrobeItemRes'
      role:
        type: string
    type: object
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.RecommendedOutfitRes:
    properties:
      explanation:
        type: string
      items:
        items:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.RecommendedItemGroup'
        type: array
      title:
        type: string
    type: object
...
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.SendChatMessageReq:
    properties:
      content:
        type: string
    required:
    - content
    type: object
...
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.WardrobeItemRes:
    properties:
      category:
        $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.CategoryRes'
      color:
        type: string
      createdAt:
        type: string
      fit:
        type: string
      id:
        type: string
      imagePublicId:
        type: string
      imageUrl:
        type: string
      isLocked:
        type: boolean
      material:
        type: string
      pattern:
        type: string
      seasonality:
        type: string
      status:
        $ref: '#/definitions/smart-wardrobe-be_internal_shared_domain_constants_wardrobestatus.WardrobeItemStatus'
      style:
        type: string
      userId:
        type: string
    type: object
...
  smart-wardrobe-be_internal_shared_domain_constants_wardrobestatus.WardrobeItemStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    format: int32
    type: integer
    x-enum-comments:
      Failed: AI processing failed
      Processing: AI processing in background
    x-enum-descriptions:
    - ""
    - ""
    - ""
    - AI processing in background
    - AI processing failed
    x-enum-varnames:
    - InWardrobe
    - Selling
    - Sold
    - Processing
    - Failed
  smart-wardrobe-be_internal_shared_presentation.APIResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: API docs for Smart Wardrobe.
  title: Smart Wardrobe API Document
  version: "1.0"
paths:
  /api/v1/ai/chat/sessions:
    get:
      consumes:
      - application/json
      description: Lấy tất cả các phiên trò chuyện của người dùng hiện tại với stylist
        AI
      produces:
      - application/json
      responses:
        "200":
          description: Lấy danh sách cuộc trò chuyện thành công
          schema:
            allOf:
            - $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.ChatSessionRes'
                  type: array
              type: object
      summary: Lấy danh sách cuộc trò chuyện AI
      tags:
      - Wardrobe AI
    post:
      consumes:
      - application/json
      description: Khởi tạo một phiên tư vấn phong cách thời trang mới với stylist
        AI
      parameters:
      - description: Yêu cầu tạo cuộc trò chuyện
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.CreateChatSessionReq'
      produces:
      - application/json
      responses:
        "201":
          description: Tạo cuộc trò chuyện thành công
          schema:
            allOf:
            - $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.ChatSessionRes'
              type: object
      summary: Tạo cuộc trò chuyện AI mới
      tags:
      - Wardrobe AI
  /api/v1/ai/chat/sessions/{contextID}/archive:
    patch:
      consumes:
      - application/json
      description: Lưu trữ/ẩn cuộc trò chuyện với stylist AI
      parameters:
      - description: ID cuộc trò chuyện
        in: path
        name: contextID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lưu trữ cuộc trò chuyện thành công
          schema:
            $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
      summary: Lưu trữ cuộc trò chuyện AI
      tags:
      - Wardrobe AI
  /api/v1/ai/chat/sessions/{contextID}/messages:
    get:
      consumes:
      - application/json
      description: Lấy toàn bộ các tin nhắn trong một phiên trò chuyện cụ thể
      parameters:
      - description: ID cuộc trò chuyện
        in: path
        name: contextID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lấy lịch sử tin nhắn thành công
          schema:
            allOf:
            - $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.ChatMessageRes'
                  type: array
              type: object
      summary: Lấy lịch sử tin nhắn AI
      tags:
      - Wardrobe AI
  /api/v1/ai/chat/sessions/{contextID}/messages/stream:
    post:
      consumes:
      - application/json
      description: Gửi tin nhắn cho stylist AI và nhận phản hồi dạng stream sự kiện
        (Server-Sent Events)
      parameters:
      - description: ID cuộc trò chuyện
        in: path
        name: contextID
        required: true
        type: string
      - description: Nội dung tin nhắn gửi đi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.SendChatMessageReq'
      produces:
      - text/event-stream
      responses:
        "200":
          description: Stream Server-Sent Events (SSE) phản hồi từ AI
          schema:
            type: string
      summary: Nhắn tin với stylist AI (Stream SSE)
      tags:
      - Wardrobe AI
  /api/v1/ai/outfit-recommendations:
    post:
      consumes:
      - application/json
      description: Nhận gợi ý phối đồ từ các trang phục có sẵn trong tủ đồ của người
        dùng dựa trên dịp, thời tiết và phong cách
      parameters:
      - description: Yêu cầu gợi ý phối đồ
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.RecommendOutfitReq'
      produces:
      - application/json
      responses:
        "200":
          description: Gợi ý phối đồ thành công
          schema:
            allOf:
            - $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.RecommendedOutfitRes'
              type: object
      summary: Gợi ý phối đồ từ tủ đồ
      tags:
      - Wardrobe AI
...
swagger: "2.0"